% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_categorical_variables.R
\name{helper_summarize_cat}
\alias{helper_summarize_cat}
\title{Helper:Calculate Absolute and Relative Frequencies of Categorical Variables}
\usage{
helper_summarize_cat(
  data,
  cat_vec,
  group_var = FALSE,
  treatment_arm = FALSE,
  new_line = FALSE,
  measures_cat = c("absolute", "relative"),
  measure_style = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data for which the Table 1 should be generated}

\item{cat_vec}{A character vector specifying the the variable names for which  the summary measures should be created.}

\item{group_var}{A character string specifying the grouping variable for stratification.
If set to \code{FALSE} (default), the Table 1 is not stratified by groups.}

\item{treatment_arm}{A character string indicating the main grouping variable. Subgroups defined by group_var are nested within each level of this variable.
For example, in a dataset with three treatment arms ("treatment": A, B, C) and two subgroups within each
treatment category ("sub_group": I, II), set \code{group_var} to "sub_group" and \code{treatment_arm} to "treatment".
If p-values or SMD options are selected, the subgroups within the treatment groups are compared.
Default is \code{FALSE} if no treatment arms are used.}

\item{new_line}{logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, an empty row is added after every variable
in the table, which can improve clarity, especially when returning the Table 1 as a flextable object.}

\item{measures_cat}{A character vector indicating how categorical variables should be summarized. Options are
\code{"absolute"}, \code{"relative"}, or both. Default is \code{c("absolute", "relative")} for displaying both absolute and
relative frequencies.}

\item{measure_style}{A logical value (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, the function \code{cat_unify_names}
is called to  combine absolute and relative frequencies into  a single column.}
}
\value{
A \code{data.frame} containing the summary measures of the chosen variables.
}
\description{
This function is primarily used internally by \code{Table1_flex()}.
It calculates summary statistics for categorical variables.
One can choose to display only absolute frequencies, relative frequencies, or both.
It works with one or more groups and also supports nested group comparisons for the comparison of subgroup statistics within a main group.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  # Load pbc data from the survival package
  pbc <- survival::pbc

  baseline_var <- c("age", "chol", "platelet")

  pbc[c("stage", "trt", "edema", "hepato")] <-
    lapply(pbc[c("stage", "trt", "edema", "hepato")], as.factor)

  pbc <- pbc[!is.na(pbc$trt), ]

  pbc$trt <- ifelse(pbc$trt == "1",
    "D-penicillamine",
    "Placebo"
  )

  # Categorical summary measure for one group -------------------------------------------

  helper_summarize_cat(
    data = pbc,
    cat_vec = c("stage", "edema", "hepato"),
    group_var = FALSE,
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_cat = c("absolute", "relative"),
    measure_style = FALSE
  )

  # Categorical summary measures summarized into one column
  helper_summarize_cat(
    data = pbc,
    cat_vec = c("stage", "edema", "hepato"),
    group_var = FALSE,
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_cat = c("absolute", "relative"),
    measure_style = TRUE
  )

  # Categorical summary measures for two groups -----------------------------------------

  helper_summarize_cat(
    data = pbc,
    cat_vec = c("stage", "edema", "hepato"),
    group_var = "trt",
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_cat = c("absolute", "relative"),
    measure_style = TRUE
  )


  # Categorical summary measures for a nested group structure ----------------------------

  helper_summarize_cat(
    data = pbc,
    cat_vec = c("stage", "edema", "hepato"),
    group_var = "sex",
    treatment_arm = "trt",
    new_line = FALSE,
    measures_cat = c("absolute", "relative"),
    measure_style = TRUE
  )
}

}
\keyword{internal}
