% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_missing_values.R
\name{helper_summarize_missings}
\alias{helper_summarize_missings}
\title{Helper: Calculate Absolute and Relative Frequency of Missing Values}
\usage{
helper_summarize_missings(
  data,
  var_vec,
  group_var = FALSE,
  treatment_arm = FALSE,
  measures = c("absolute", "relative"),
  measure_style = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data for which the Table 1 should be generated}

\item{var_vec}{A character vector specifying  the variable names for which the Table 1 summary should be created.}

\item{group_var}{A character string specifying the grouping variable for stratification.
If set to \code{FALSE} (default), the Table 1 is not stratified by groups.}

\item{treatment_arm}{A character string indicating the main grouping variable. Subgroups defined by group_var are nested within each level of this variable.
For example, in a dataset with three treatment arms ("treatment": A, B, C) and two subgroups within each
treatment category ("sub_group": I, II), set \code{group_var} to "sub_group" and \code{treatment_arm} to "treatment".
If p-values or SMD options are selected, the subgroups within the treatment groups are compared.
Default is \code{FALSE} if no treatment arms are used.}

\item{measures}{A character vector indicating how categorical variables should be summarized. Options are
\code{"absolute"}, \code{"relative"}, or both. Default is \code{c("absolute", "relative")} for displaying both absolute and
relative frequencies.}

\item{measure_style}{A logical value (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, the function \code{cat_unify_names} will be called to  combine absolute and relative frequencies into
a single column.}
}
\value{
A \code{data.frame} containing the summary measures for missing values specified in the input parameters.
}
\description{
This function is primarily used internally by \code{Table1_flex()}.
It calculates summary statistics for missing values in variables.
One can choose to display only absolute frequencies, relative frequencies, or both.
It works with one or more groups and also supports nested group comparisons  for the comparison of missing values between subgroups
within an main group.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  # Load pbc data from the survival package
  pbc <- survival::pbc

  baseline_var <- c("age", "chol", "platelet")

  pbc[c("stage", "trt", "edema", "hepato")] <-
    lapply(pbc[c("stage", "trt", "edema", "hepato")], as.factor)

  pbc <- pbc[!is.na(pbc$trt), ]

  pbc$trt <- ifelse(pbc$trt == "1",
    "D-penicillamine",
    "Placebo"
  )

  # Missing summary measure for one group -------------------------------------------

  helper_summarize_missings(
    data = pbc,
    var_vec = baseline_var,
    group_var = FALSE,
    treatment_arm = FALSE,
    measures = c("absolute", "relative"),
    measure_style = FALSE
  )

  # Missing summary measure summarized into one column
  helper_summarize_missings(
    data = pbc,
    var_vec = baseline_var,
    group_var = FALSE,
    treatment_arm = FALSE,
    measures = c("absolute", "relative"),
    measure_style = TRUE
  )

  # Missing summary measure for two groups -----------------------------------------

  helper_summarize_missings(
    data = pbc,
    var_vec = baseline_var,
    group_var = "trt",
    treatment_arm = FALSE,
    measures = c("absolute", "relative"),
    measure_style = TRUE
  )


  # Missing summary measure for a nested group structure ----------------------------

  helper_summarize_missings(
    data = pbc,
    var_vec = baseline_var,
    group_var = "sex",
    treatment_arm = "trt",
    measures = c("absolute", "relative"),
    measure_style = TRUE
  )
}

}
\keyword{internal}
