% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_numeric_variables.R
\name{helper_summarize_num}
\alias{helper_summarize_num}
\title{Helper function: Calculates Summary Statistics for Multiple Numeric Variables}
\usage{
helper_summarize_num(
  data,
  num_vec,
  group_var = FALSE,
  treatment_arm = FALSE,
  new_line = FALSE,
  measures_num = c("median", "min", "max"),
  measure_style = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data for which the Table 1 should be generated}

\item{num_vec}{A character vector specifying the variable names for which
the summary measures should be created.}

\item{group_var}{A character string specifying the grouping variable for stratification.
If set to \code{FALSE} (default), the Table 1 is not stratified by groups.}

\item{treatment_arm}{A character string indicating the main grouping variable. Subgroups defined by group_var are nested within each level of this variable.
For example, in a dataset with three treatment arms ("treatment": A, B, C) and two subgroups within each
treatment category ("sub_group": I, II), set \code{group_var} to "sub_group" and \code{treatment_arm} to "treatment".
If p-values or SMD options are selected, the subgroups within the treatment groups are compared.
Default is \code{FALSE} if no treatment arms are used.}

\item{new_line}{logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, an empty row is added after every variable
in the table, which can improve clarity, especially when returning the Table 1 as a flextable object.}

\item{measures_num}{A character vector defining the summary statistics for numeric variables. Options include \code{"median"},
\code{"min"}, \code{"max"}, \code{"quantile1"} (25th percentile), \code{"quantile3"} (75th percentile), \code{"mean"}, and \code{"sd"}
(standard deviation).  The number of measures is limited to three (e.g., mean(sd), median(min-max), or
median(quantile1-quantile3)).}

\item{measure_style}{A logical value (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, the helper function \code{cat_unify_names}
will be called to combine the selected summary measures into a single column.}
}
\value{
A \code{data.frame} containing the summary measures specified in the input parameters.
}
\description{
This function is primarily used internally by \code{Table1_flex()}.
It calculates summary statistics for numerical variables.
One can choose to display the median, mean, standard deviation, first and third quartiles, minimum, maximum, or a combination of these.
It works with one or more groups and nested group comparisons for the comparison of subgroup statistics within a main group.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  # Load pbc data from the survival package
  pbc <- survival::pbc

  baseline_var <- c("age", "chol", "platelet")

  pbc[c("stage", "trt", "edema", "hepato")] <-
    lapply(pbc[c("stage", "trt", "edema", "hepato")], as.factor)

  pbc <- pbc[!is.na(pbc$trt), ]

  pbc$trt <- ifelse(pbc$trt == "1",
    "D-penicillamine",
    "Placebo"
  )

  # Numerical summary measure for one group -------------------------------------------

  helper_summarize_num(
    data = pbc,
    num_vec = c("age", "chol", "platelet"),
    group_var = FALSE,
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_num = c("median", "quantile1", "quantile3"),
    measure_style = FALSE
  )

  # Numerical summary measure summarized into one column
  helper_summarize_num(
    data = pbc,
    num_vec = c("age", "chol", "platelet"),
    group_var = FALSE,
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_num = c("median", "quantile1", "quantile3"),
    measure_style = TRUE
  )

  # Numerical summary measure for two groups -----------------------------------------

  helper_summarize_num(
    data = pbc,
    num_vec = c("age", "chol", "platelet"),
    group_var = "trt",
    treatment_arm = FALSE,
    new_line = FALSE,
    measures_num = c("median", "quantile1", "quantile3"),
    measure_style = TRUE
  )


  # Numerical summary measure for a nested group structure ----------------------------


  helper_summarize_num(
    data = pbc,
    num_vec = c("age", "chol", "platelet"),
    group_var = "sex",
    treatment_arm = "trt",
    new_line = FALSE,
    measures_num = c("mean", "sd"),
    measure_style = TRUE
  )
}

}
\keyword{internal}
