% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Tab1.R
\name{Table1_flex}
\alias{Table1_flex}
\title{Generate Table 1 with Customizable Summary Measures}
\usage{
Table1_flex(
  data,
  variables,
  group_var = FALSE,
  treatment_arm = FALSE,
  new_line = FALSE,
  measures_cat = c("absolute", "relative"),
  measures_num = c("median", "min", "max"),
  display_pvalue = FALSE,
  display_smd = FALSE,
  display_missings = TRUE,
  flextable_output = TRUE,
  sort_rows = NULL,
  add_measure_ident = TRUE,
  treatment_order = NULL,
  group_order = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data for which the Table 1 should be generated}

\item{variables}{A character vector specifying  the variable names for which the Table 1 summary should be created.}

\item{group_var}{A character string specifying the grouping variable for stratification.
If set to \code{FALSE} (default), the Table 1 is not stratified by groups.}

\item{treatment_arm}{A character string indicating the main grouping variable. Subgroups defined by group_var are nested within each level of this variable.
For example, in a dataset with three treatment arms ("treatment": A, B, C) and two subgroups within each
treatment category ("sub_group": I, II), set \code{group_var} to "sub_group" and \code{treatment_arm} to "treatment".
If p-values or SMD options are selected, the subgroups within the treatment groups are compared.
Default is \code{FALSE} if no treatment arms are used.}

\item{new_line}{logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, an empty row is added after every variable
in the table, which can improve clarity, especially when returning the Table 1 as a flextable object.}

\item{measures_cat}{A character vector indicating how categorical variables should be summarized. Options are
\code{"absolute"}, \code{"relative"}, or both. Default is \code{c("absolute", "relative")} for displaying both absolute and
relative frequencies.}

\item{measures_num}{A character vector defining the summary statistics for numeric variables. Options include \code{"median"},
\code{"min"}, \code{"max"}, \code{"quantile1"} (25th percentile), \code{"quantile3"} (75th percentile), \code{"mean"}, and \code{"sd"}
(standard deviation).  The number of measures is limited to three (e.g., mean(sd), median(min-max), or
median(quantile1-quantile3)).}

\item{display_pvalue}{logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, p-values are displayed.}

\item{display_smd}{logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, standardized mean differences (SMD)
is displayed.}

\item{display_missings}{A logical (TRUE/FALSE), default is \code{FALSE}. If \code{TRUE}, the frequency of missing values for
each variable is shown in the output.}

\item{flextable_output}{A logical (TRUE/FALSE), default is \code{TRUE}. If \code{TRUE}, the Table 1 is returned as a
\code{flextable} object. If \code{FALSE}, it is returned as a \code{data.frame}.}

\item{sort_rows}{A character vector specifying the row order in which the variables should appear in the output table.
The default is NULL, which sorts the variables alphabetically.}

\item{add_measure_ident}{A logical (TRUE/FALSE), default is \code{TRUE}. If TRUE, the summary measures identifiers, e.g. median (min-max),
mean (sd), n (\%) is added to the Table 1.}

\item{treatment_order}{character vector containing the categories of the \code{treatment_arm} parameter in the desired column order in the output table.}

\item{group_order}{character vector containing the categories of the \code{group_var} parameter in the desired column order in the output table.}
}
\value{
A \code{data.frame} or \code{flextable} object, depending on the selected output option.
The returned Table 1 contains the summary measures specified in the input parameters,
including any chosen statistics, p-values, standardized mean differences (SMD), and missing value frequencies.
}
\description{
This function creates a highly flexible Table 1 for descriptive statistics.
It allows users to customize the output, choose summary measures, display missing values,
p-values (testing for difference between two groups),and standardized mean differences (SMD).
}
\details{
Testing for differences in variables between groups.
For numerical variables, a Wilcoxon rank-sum test is used to assess differences between the median of  two groups at a significance level of 5\\%.
For categorical variables, a Chi-square test is used to assess differences between groups at a significance level of 5\\%. If the expected
absolute count is equal or below 5, then the Fisher's exact test is used instead.
Till now, only group comparisons between two groups is implemented.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  # Load pbc data from the survival package
  pbc <- survival::pbc

  baseline_var <- c("age", "chol", "sex", "stage", "platelet")

  pbc[c("stage", "trt", "edema", "hepato")] <-
    lapply(pbc[c("stage", "trt", "edema", "hepato")], as.factor)

  pbc <- pbc[!is.na(pbc$trt), ]
pbc$trt <- ifelse(pbc$trt == "1",
"D-penicillamine",
"Placebo")

# Table 1 for one group -------------------------------------------

tab1_ex <- Table1_flex(
data = pbc,
variables = baseline_var
)
flextable::autofit(tab1_ex)

# choosing other summary measures
Table1_flex(
data = pbc,
variables = baseline_var,
measures_num = c("mean", "sd"),
measures_cat = "relative"
)

# Returning a data.frame for further processing
# and other summary measures
Table1_flex(
data = pbc,
variables = baseline_var,
measures_num = c("mean", "sd"),
measures_cat = c("absolute","relative"),
flextable_output = FALSE
)

# Table 1 for two groups -----------------------------------------

tab1_ex <- Table1_flex(
data = pbc,
variables = baseline_var,
group_var = "trt",
add_measure_ident = FALSE,
new_line = TRUE,
sort_rows = c("age", "sex", "stage", "chol", "platelet"),
group_order = c("Placebo", "D-penicillamine")
)

flextable::autofit(tab1_ex)

# with additionally p-values and standarized mean differences
tab1_ex <- Table1_flex(
data = pbc,
variables = baseline_var,
group_var = "trt",
add_measure_ident = FALSE,
new_line = TRUE,
sort_rows = c("age", "sex", "stage", "chol", "platelet"),
group_order = c("Placebo", "D-penicillamine"),
display_pvalue = TRUE,
display_smd = TRUE
)

flextable::autofit(tab1_ex)


# Table 1 for a nested group structure ----------------------------

baseline_var <- c("age", "chol", "platelet", "stage")
tab1_ex <- Table1_flex(
data = pbc,
variables = baseline_var,
group_var = "sex",
treatment_arm = "trt",
add_measure_ident = TRUE,
sort_rows = c("age", "stage", "chol", "platelet"),
flextable_output = TRUE,
display_pvalue = TRUE
)

flextable::autofit(tab1_ex)
}

}
\references{
Austin, P. C. (2011). An Introduction to Propensity Score Methods for Reducing the Effects of Confounding in Observational Studies. Multivariate Behavioral Research, 46(3), 399–424. https://doi.org/10.1080/00273171.2011.568786
}
